+++ transformers/struct_test.go
@@ -6,6 +6,7 @@ import (
 	"slices"
 	"testing"
 	"time"
+	"unsafe"
 
 	"github.com/apache/arrow/go/v17/arrow"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
@@ -476,6 +477,216 @@ func TestTableFromGoStruct(t *testing.T) {
 
 			if diff := cmp.Diff(table.PrimaryKeys(), tt.want.PrimaryKeys()); diff != "" {
 				t.Fatalf("table does not match expected. diff (-got, +want): %v", diff)
+			}
+		})
+	}
+}
+
+func TestJSONTypeSchema(t *testing.T) {
+	tests := []struct {
+		name       string
+		testStruct any
+		maxDepth   int
+		want       map[string]string
+	}{
+		{
+			name: "simple map",
+			testStruct: struct {
+				Tags map[string]string `json:"tags"`
+			}{},
+			want: map[string]string{
+				"tags": `{"utf8":"utf8"}`,
+			},
+		},
+		{
+			name: "simple map pointer",
+			testStruct: struct {
+				Tags *map[string]string `json:"tags"`
+			}{},
+			want: map[string]string{
+				"tags": `{"utf8":"utf8"}`,
+			},
+		},
+		{
+			name: "simple array",
+			testStruct: struct {
+				Items []struct {
+					Name string `json:"name"`
+				} `json:"items"`
+			}{},
+			want: map[string]string{
+				"items": `[{"name":"utf8"}]`,
+			},
+		},
+		{
+			name: "simple array pointer",
+			testStruct: struct {
+				Items *[]struct {
+					Name string `json:"name"`
+				} `json:"items"`
+			}{},
+			want: map[string]string{
+				"items": `[{"name":"utf8"}]`,
+			},
+		},
+		{
+			name: "simple struct",
+			testStruct: struct {
+				Item struct {
+					Name string `json:"name"`
+				} `json:"item"`
+			}{},
+			want: map[string]string{
+				"item": `{"name":"utf8"}`,
+			},
+		},
+		{
+			name: "complex struct",
+			testStruct: struct {
+				Item struct {
+					Name         string            `json:"name"`
+					Tags         map[string]string `json:"tags"`
+					FlatItems    []string          `json:"flat_items"`
+					ComplexItems []struct {
+						Name string `json:"name"`
+					} `json:"complex_items"`
+				} `json:"item"`
+			}{},
+			want: map[string]string{
+				"item": `{"complex_items":[{"name":"utf8"}],"flat_items":["utf8"],"name":"utf8","tags":{"utf8":"utf8"}}`,
+			},
+		},
+		{
+			name: "multiple json columns",
+			testStruct: struct {
+				Tags map[string]string `json:"tags"`
+				Item struct {
+					Name         string            `json:"name"`
+					Tags         map[string]string `json:"tags"`
+					FlatItems    []string          `json:"flat_items"`
+					ComplexItems []struct {
+						Name string `json:"name"`
+					} `json:"complex_items"`
+				} `json:"item"`
+			}{},
+			want: map[string]string{
+				"item": `{"complex_items":[{"name":"utf8"}],"flat_items":["utf8"],"name":"utf8","tags":{"utf8":"utf8"}}`,
+			},
+		},
+		{
+			name: "handles any type in struct",
+			testStruct: struct {
+				Item struct {
+					Name   string `json:"name"`
+					Object any    `json:"object"`
+				} `json:"item"`
+			}{},
+			want: map[string]string{
+				"item": `{"name":"utf8","object":"any"}`,
+			},
+		},
+		{
+			name: "handles map from string to any",
+			testStruct: struct {
+				Tags map[string]any `json:"tags"`
+			}{},
+			want: map[string]string{
+				"tags": `{"utf8":"any"}`,
+			},
+		},
+		{
+			name: "handles array of any",
+			testStruct: struct {
+				Items []any `json:"items"`
+			}{},
+			want: map[string]string{
+				"items": `["any"]`,
+			},
+		},
+		{
+			name: "stops at the default depth of 5",
+			testStruct: struct {
+				Level0 struct {
+					Level1 struct {
+						Level2 struct {
+							Level3 struct {
+								Level4 struct {
+									Level5 struct {
+										Level6 struct {
+											Name string `json:"name"`
+										} `json:"level6"`
+									} `json:"level5"`
+								} `json:"level4"`
+							} `json:"level3"`
+						} `json:"level2"`
+					} `json:"level1"`
+				} `json:"level0"`
+			}{},
+			want: map[string]string{
+				"level0": `{"level1":{"level2":{"level3":{"level4":{"level5":"json"}}}}}`,
+			},
+		},
+		{
+			name:     "stops at the configured depth",
+			maxDepth: 2,
+			testStruct: struct {
+				Level0 struct {
+					Level1 struct {
+						Level2 struct {
+							Level3 struct {
+								Level4 struct {
+									Level5 struct {
+										Level6 struct {
+											Name string `json:"name"`
+										} `json:"level6"`
+									} `json:"level5"`
+								} `json:"level4"`
+							} `json:"level3"`
+						} `json:"level2"`
+					} `json:"level1"`
+				} `json:"level0"`
+			}{},
+			want: map[string]string{
+				"level0": `{"level1":{"level2":{"level3":"json"}}}`,
+			},
+		},
+		{
+			name: "ignores non exported and ignored types",
+			testStruct: struct {
+				Item struct {
+					nonExported   string
+					f             func()
+					c             chan int
+					unsafePointer unsafe.Pointer
+					Exported      string `json:"exported"`
+				} `json:"item"`
+			}{},
+			want: map[string]string{
+				"item": `{"exported":"utf8"}`,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			table := schema.Table{
+				Name: "test",
+			}
+			opts := []StructTransformerOption{}
+			if tt.maxDepth > 0 {
+				opts = append(opts, WithMaxJSONTypeSchemaDepth(tt.maxDepth))
+			}
+			transformer := TransformWithStruct(tt.testStruct, opts...)
+			err := transformer(&table)
+			if err != nil {
+				t.Fatal(err)
+			}
+			for col, schema := range tt.want {
+				column := table.Column(col)
+				if diff := cmp.Diff(column.TypeSchema, schema); diff != "" {
+					t.Fatalf("table does not match expected. diff (-got, +want): %v", diff)
+				}
 			}
 		})
 	}